# Conflicts:
#	API/db/exercise.sqlite3
#	fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/FIITNavHost.kt
#	fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/ui/viewmodels/HomeViewModel.kt
#	fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/util/ViewModelFactory.kt
#	fiit-mobile/app/src/main/res/values-en/strings.xml
#	fiit-mobile/app/src/main/res/values/strings.xml

+++ fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/data/network/model/exercises/NetworkExerciseContent.kt
@@ -9,7 +9,7 @@ data class NetworkExerciseContent (
     @SerializedName("name"     ) var name     : String,
     @SerializedName("detail"   ) var detail   : String,
     @SerializedName("type"     ) var type     : String,
-    @SerializedName("date"     ) var date     : Int?    = null,
+    @SerializedName("date"     ) var date     : Long?    = null,
     @SerializedName("metadata" ) var imgUrl : String? = null
 
 )
